Refactor celery paths, add conn_max_age
