+++ .docker/scraper/start-django.sh
@@ -1,3 +1,3 @@
 #!/usr/bin/env bash
 
-celery -A trquake.celery.app worker -B --concurrency=15 -l INFO
+celery -A trquake.celery_service.app worker -B --concurrency=15 -l INFO
