+++ docker-compose.yml
@@ -43,7 +43,7 @@ services:
     <<: *django
     container_name: trquake-workers
     ports: []
-    command: watchmedo auto-restart --directory=/app --pattern=*.py --recursive -- celery -A trquake.celery.app worker -B --concurrency=5 -l DEBUG
+    command: watchmedo auto-restart --directory=/app --pattern=*.py --recursive -- celery -A trquake.celery_service.app worker -B --concurrency=5 -l DEBUG
     environment:
       - DJANGO_SETTINGS_MODULE=trquake.settings.development
       - POSTGRES_PASSWORD=debug
