+++ trquake/celery_service.py
@@ -1,6 +1,5 @@
 # Third Party
 from celery import Celery
-from celery.schedules import crontab
 
 app = Celery("trquake")
